+++ .vscode/settings.json
@@ -6,5 +6,8 @@
 	"editor.defaultFormatter": "esbenp.prettier-vscode",
 	"[typescript]": {
 		"editor.defaultFormatter": "esbenp.prettier-vscode"
-	}
+	},
+	"cSpell.words": [
+		"Pluggable"
+	]
 }
