+++ packages/auth/src/Providers/CognitoProvider/types/machines/authorizationMachine.ts
@@ -14,7 +14,7 @@
 import { CognitoProviderConfig } from '../../CognitoProvider';
 import { ActorRefFrom } from 'xstate';
 import { fetchAuthSessionStateMachine } from '../../machines/fetchAuthSessionStateMachine';
-import { CognitoService } from '../../serviceClass';
+import { CognitoService } from '../../services/CognitoService';
 
 // info/context needed to pass to the fetchAuthSessionStateMachine in order to perform the fetch Session
 // First, fetch user pool tokens (JWT) from the user pool
@@ -33,16 +33,16 @@ export type FetchAuthSessionActorRef = ActorRefFrom<
 
 export interface AuthorizationMachineContext {
 	actorRef?: FetchAuthSessionActorRef;
-	config: null | CognitoProviderConfig;
-	service: null | CognitoService;
-	userPoolTokens: null | {
+	config?: null | CognitoProviderConfig;
+	service?: null | CognitoService;
+	userPoolTokens?: null | {
 		idToken: string;
 		accessToken: string;
 		refreshToken: string;
 	};
-	identityID: null;
-	AWSCredentials: null;
-	sessionInfo: any;
+	identityID?: null;
+	AWSCredentials?: null;
+	sessionInfo?: any;
 	clientConfig?: CognitoProviderConfig;
 }
