+++ packages/auth/src/Providers/CognitoProvider/types/machines/refreshSessionMachine.ts
@@ -13,7 +13,7 @@
 
 import { UserPoolTokens } from './authorizationMachine';
 import { CognitoProviderConfig } from '../../CognitoProvider';
-import { CognitoService } from '../../serviceClass';
+import { CognitoService } from '../../services/CognitoService';
 
 export interface RefreshSessionStateMachineContext {
 	config: null | CognitoProviderConfig;
