+++ packages/auth/src/Providers/CognitoProvider/types/model/index.ts
@@ -1,0 +1,2 @@
+export * from './signin';
+export * from './session';
