+++ packages/auth/src/Providers/CognitoProvider/types/model/session/AWSCredentials.ts
@@ -1,0 +1,6 @@
+export type AWSCredentials = {
+	accessKeyId: string;
+	secretAccessKey: string;
+	sessionToken?: string;
+	expiration: Date;
+};
