+++ packages/auth/src/Providers/CognitoProvider/types/model/session/CognitoSessionData.ts
@@ -1,0 +1,7 @@
+// TODO: Rename
+export type CognitoSessionData = {
+	accessToken: string;
+	idToken: string;
+	refreshToken: string;
+	expiration: number;
+};
