+++ packages/auth/src/Providers/CognitoProvider/types/model/session/index.ts
@@ -1,0 +1,1 @@
+export { CognitoSessionData } from './CognitoSessionData';
