+++ packages/auth/src/Providers/CognitoProvider/types/model/signin/CognitoCompletePasswordOptions.ts
@@ -1,0 +1,6 @@
+export type CognitoCompletePasswordOptions = {
+	username: string;
+	newPassword: string;
+	requiredAttributes?: { [key: string]: any };
+	session: string;
+};
