+++ packages/auth/src/Providers/CognitoProvider/types/model/signin/CognitoConfirmSignInOptions.ts
@@ -1,0 +1,7 @@
+import { ConfirmSignInParams } from 'src/types/model/signin/ConfirmSignInParams';
+
+export type CognitoConfirmSignInOptions = ConfirmSignInParams & {
+	session: string;
+	username: string;
+	clientId: string;
+};
