+++ packages/auth/src/Providers/CognitoProvider/types/model/signin/index.ts
@@ -1,0 +1,2 @@
+export { CognitoConfirmSignInOptions } from './CognitoConfirmSignInOptions';
+export { CognitoCompletePasswordOptions } from './CognitoCompletePasswordOptions';
