+++ packages/auth/src/Providers/CognitoProvider/types/model/signup/CognitoSignUpParams.ts
@@ -1,0 +1,6 @@
+import { SignUpParams } from 'src/types/model/signup/SignUpParams';
+
+export type CognitoSignUpParams = SignUpParams & {
+	validationData?: { [key: string]: any };
+	clientMetadata?: { [key: string]: string };
+};
