+++ packages/auth/src/Providers/CognitoProvider/types/model/user/CognitoUser.ts
@@ -1,0 +1,16 @@
+import { AmplifyUser } from 'src/types/model/user/AmplifyUser';
+import { AWSCredentials } from '../session/AWSCredentials';
+
+export type CognitoUser = AmplifyUser & {
+	credentials?: {
+		// scope
+		[key: string]: {
+			jwt?: {
+				idToken: string;
+				accessToken: string;
+				refreshToken: string;
+			};
+			aws?: AWSCredentials;
+		};
+	};
+};
