+++ packages/auth/src/types/model/index.ts
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
+ * the License. A copy of the License is located at
+ *
+ *     http://aws.amazon.com/apache2.0/
+ *
+ * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
+export { ConfirmSignInPluginOptions } from './signin/ConfirmSignInPluginOptions';
+export { ConfirmSignInParams } from './signin/ConfirmSignInParams';
+export { SignInResult } from './signin/SignInResult';
+
+export { ConfirmSignUpParams } from './signup/ConfirmSignUpParams';
+export { ConfirmSignUpResult } from './signup/ConfirmSignUpResult';
+export { SignUpParams } from './signup/SignUpParams';
+export { SignUpPluginOptions } from './signup/SignUpPluginOptions';
+export { SignUpResult } from './signup/SignUpResult';
+
+export { AmplifyUser } from './user/AmplifyUser';
