+++ packages/auth/src/types/model/signin/ConfirmSignInPluginOptions.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
  * the License. A copy of the License is located at
@@ -11,7 +11,6 @@
  * and limitations under the License.
  */
 
-export * from './Auth';
-export * from './AuthPluggable';
-export * from './Provider';
-export * from './model/index';
+import { AuthPluginOptions } from '../AuthPluginOptions';
+
+export type ConfirmSignInPluginOptions = AuthPluginOptions & {};
